To-do-list_MSWD
To-do-list_MSWD 
This web app allows you to create a list of to-dos with custom categories. The to-dos get synched to a database under your user profile. You can leave the site or logout at anytime, then log back in and your to-dos will still be there! 
 
Made this project for me to better learn React.js with the Flux pattern, as well as handling user authentication using social login APIs. Built on node.js/express/mongoDB on the back end, and React.js/Flux on the front end. 
 
Running the project 
1.Install node.js and npm 
2.https://github.com/goutam29/To-Do-List_MSWD_Project
3.Install dependencies $ npm install 
4.Start app $ npm start 
5.Go to http://localhost:3000